backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
websockets
numpy
sqlalchemy
requests
openai
python-dotenv
